* origin/main:
  docs: Updated changelog
  test: Add some additional test coverage
  fix: Typo in filename
  docs: Mention downloading the reference and sample data for modelling centers
  refactor: Rename to DatasetRegistryManager
  test: Rework the cli tests
  test: Add tests for the dataset registry
  fix: Rename ilamb data registry
  chore: rename registry
  feat: Make it more plugin orientated
  chore: changelog
  chore: Add the option to fetch the annual cycle datasets
  feat: Update to support providing a registry file

+++ packages/ref/src/cmip_ref/cli/datasets.py
@@ -16,9 +16,10 @@ from rich.console import Console
 from cmip_ref.cli._utils import pretty_print_df
 from cmip_ref.datasets import get_dataset_adapter
 from cmip_ref.models import Dataset
+from cmip_ref.provider_registry import ProviderRegistry
 from cmip_ref.solver import solve_metrics
-from cmip_ref.testing import SAMPLE_DATA_VERSION, fetch_sample_data
-from cmip_ref_core.dataset_registry import build_reference_data_registry, fetch_all_files
+from cmip_ref.testing import fetch_sample_data
+from cmip_ref_core.dataset_registry import dataset_registry_manager, fetch_all_files
 from cmip_ref_core.datasets import SourceDatasetType
 
 app = typer.Typer(help=__doc__)
@@ -151,11 +152,6 @@ def ingest(  # noqa: PLR0913
 
 @app.command(name="fetch-sample-data")
 def _fetch_sample_data(
-    version: Annotated[
-        str,
-        "The version tag of the sample data to fetch. "
-        "Defaults to the current version of data expected by the test suite",
-    ] = SAMPLE_DATA_VERSION,
     force_cleanup: Annotated[bool, typer.Option(help="If True, remove any existing files")] = False,
     symlink: Annotated[
         bool, typer.Option(help="If True, symlink files into the output directory, otherwise perform a copy")
@@ -168,27 +164,42 @@ def _fetch_sample_data(
     This operation may fail if the test data directory does not exist,
     as is the case for non-source-based installations.
     """
-    logger.info(f"Fetching data for version {version}")
-    fetch_sample_data(version=version, force_cleanup=force_cleanup, symlink=symlink)
+    fetch_sample_data(force_cleanup=force_cleanup, symlink=symlink)
 
 
-@app.command(name="fetch-obs4ref-data")
-def fetch_obs4ref_data(
-    output_directory: Annotated[Path, typer.Option(help="Output directory where files will be saved")],
+@app.command(name="fetch-data")
+def fetch_data(
+    ctx: typer.Context,
+    registry: Annotated[str, typer.Option(help="Name of the data registry to use")],
+    output_directory: Annotated[
+        Path | None, typer.Option(help="Output directory where files will be saved")
+    ] = None,
     force_cleanup: Annotated[bool, typer.Option(help="If True, remove any existing files")] = False,
     symlink: Annotated[
         bool, typer.Option(help="If True, symlink files into the output directory, otherwise perform a copy")
     ] = False,
 ) -> None:
     """
-    Fetch non-published Obs4MIPs data that is used by the REF
+    Fetch REF-specific datasets
 
     These datasets have been verified to have open licenses
     and are in the process of being added to Obs4MIPs.
     """
-    if force_cleanup and output_directory.exists():
+    config = ctx.obj.config
+    db = ctx.obj.database
+
+    # Setup the provider registry to register any dataset registries in the configured providers
+    ProviderRegistry.build_from_config(config, db)
+
+    if output_directory and force_cleanup and output_directory.exists():
         logger.warning(f"Removing existing directory {output_directory}")
         shutil.rmtree(output_directory)
 
-    data_registry = build_reference_data_registry()
-    fetch_all_files(data_registry, output_directory, symlink=symlink)
+    try:
+        _registry = dataset_registry_manager[registry]
+    except KeyError:
+        logger.error(f"Registry {registry} not found")
+        logger.error(f"Available registries: {', '.join(dataset_registry_manager.keys())}")
+        raise typer.Exit(code=1)
+
+    fetch_all_files(_registry, output_directory, symlink=symlink)
